fix(typing): declare model/RAG contracts and pin third-party stub boundaries
mypy phase 6: correct definition-layer contracts so downstream errors resolve at the source - FontScheme sizes and slide body font size are float not int, Role enum keys, PromptConfig nullable columns, BaseModelClient vision client class, RetryConfig default factory, apply_to_outline returns PPTOutline, and VectorStore TF-IDF narrowing accessor. Add specific-code ignores only for genuine stub gaps: fcntl flock (POSIX-only), pywin32 services, lxml, sentence-transformers optional dependency, and a socket.getaddrinfo address cast.

+++ office_agent/ppt_agent/models.py
@@ -73,10 +73,10 @@ class FontScheme:
     title_en: str = "Calibri"
     body_cn: str = "微软雅黑"
     body_en: str = "Calibri"
-    title_size: int = 28
-    subtitle_size: int = 20
-    body_size: int = 18
-    caption_size: int = 14
+    title_size: float = 28.0
+    subtitle_size: float = 20.0
+    body_size: float = 18.0
+    caption_size: float = 14.0
 
 
 @dataclass
@@ -103,7 +103,7 @@ class SlideContent:
     quote_source: str = ""          # 引用来源
     timeline_items: list = field(default_factory=list)  # [(time, title, desc)]
     notes: str = ""                 # 备注
-    body_font_size: Optional[int] = None  # 内容密度自适应字号
+    body_font_size: Optional[float] = None  # 内容密度自适应字号
     page_number: int = 0
 
     def __post_init__(self):
